Fix local_key update to not be thwarted by retries
The base exception handler's retry loop didn't reraise when the count
was exhausted, so local_key was never updated even if a JSON decode
error was raised.

After trying to reset the device, we explicitly catch JSONDecodeErrors
(without stringification) and rekey. In some circumstances a bad key can
also result in a UnicodeDecodeError, so catch that too.

This makes things much more robust across device rebootstraps.